Reformatted conditional rendering to be based off of html rather than chromes api. Moved button handlers to their own file
