+++ background.js
@@ -1,4 +1,4 @@
-chrome.storage.sync.get("user_token", function (storage_data) {
+/*chrome.storage.sync.get("user_token", function (storage_data) {
   //BUG: user has to double click
   chrome.browserAction.onClicked.addListener(function () {
     if (storage_data.user_token) {
@@ -8,3 +8,4 @@ chrome.storage.sync.get("user_token", function (storage_data) {
     }
   });
 });
+*/
