+++ manifest.json
@@ -4,20 +4,30 @@
   "version": "1.0",
   "manifest_version": 2,
   "background": {
-    "scripts": ["background.js"],
+    "scripts": [
+      "background.js"
+    ],
     "persistent": false
   },
   "content_scripts": [
     {
-      "matches": ["<all_urls>"],
-      "js": ["contents.js"]
+      "matches": [
+        "<all_urls>"
+      ],
+      "js": [
+        "contents.js"
+      ]
     }
   ],
   "browser_action": {
     "default_icon": "kanify.png",
-    "default_title": "Kanify"
+    "default_title": "Kanify",
+    "default_popup": "popup.html"
   },
-  "permissions": ["storage", "tabs"],
+  "permissions": [
+    "storage",
+    "tabs"
+  ],
   "commands": {
     "_execute_browser_action": {
       "suggested_key": {
