+++ popup.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Kanify</title>
+  <style>
+    body {
+      font-size: 14px;
+      min-height: 180px;
+      width: 384px;
+    }
+
+    input[type="text"] {
+      width: 250px;
+      border-radius: 4px;
+      border-width: 1px;
+    }
+
+    .Input-Container {
+      flex: 1;
+      margin-bottom: 2.5%;
+    }
+
+    .highlight {
+      background-color: yellow;
+    }
+  </style>
+</head>
+
+<body>
+  <div id="root">
+    <h1>Kanify (v1.0.0)</h1>
+    <div class="Input-Container">
+
+    </div>
+    <script type="module" src="./popup_scripts/popup.js" charset="utf-8"></script>
+  </div>
+</body>
+
+</html>
